Moved DayView to view packages and created time model for this view
