+++ nl/hanze/CarParkSimulation/logic/Time.java
@@ -1,0 +1,10 @@
+package nl.hanze.CarParkSimulation.logic;
+
+/**
+ * Class Time
+ *
+ * @author Koen Hendriks
+ * @version 0.1 (08-04-2016)
+ */
+public class Time extends AbstractModel {
+}
